refactor(assembly): move (Kernel)Library::from_dir to Assembler::assemble_(kernel|library)_from_dir
